test: update Code tests to use react testing library
